extract table generation into a module
